+++ config/_templates/dataset/yaak.yaml
@@ -20,9 +20,9 @@ sources:
     (@=source_id@):
       index_column: meta/ImageMetadata.(@=source_id@)/frame_idx
       source:
-        _target_: rbyte.io.FfmpegFrameSource
+        _target_: rbyte.io.TorchCodecFrameSource
         path: ${data_dir}/(@=input_id@)/(@=source_id@).pii.mp4
-        resize_shorter_side: 324
+        device: cpu
     #@ end
   #@ end
